+++ front-figures/scripts/_argparse.py
@@ -26,7 +26,7 @@ import argparse
 from typing import Optional
 
 
-SKILL_VERSION = "0.21.0"
+SKILL_VERSION = "0.22.0"
 
 
 def make_parser(