+++ front-ux-laws/scripts/_argparse.py
@@ -25,7 +25,7 @@ import argparse
 from typing import Optional
 
 
-SKILL_VERSION = "0.21.0"
+SKILL_VERSION = "0.22.0"
 
 
 def make_parser(